work please

+++ web/templates/balview_form.php
@@ -70,6 +70,7 @@
 								<div class="col-sm-12 form-group">
 									<button type="submit" class="button2" >Send &rarr;</button>
 									<?php
+										if (!$is_create)
 										echo "<a class='button2' id= 'mylink' href='incviews.php?delete_data=".$id."'>DELETE</a>";
 									?>
 								</div>
